Move rasterize-pdf route into routes/microPractice.ts
