+++ backend/src/routes/microPractice.ts
@@ -6,11 +6,27 @@ import { dbStore, UserRole, Question, PracticeSchedule, MicroAssignment, Uploade
 import { getAuthUser, canAccessStudent } from '../auth';
 import { getStrandForLevel } from '../flnLevels';
 import { getOrInitPracticeSchedule, normalizeCompetencyName, calculateNextScheduleState } from '../services/practiceScheduleService';
-import { ROOT_DIR } from '../config';
+import { ROOT_DIR, PYTHON_BIN, AI_SERVICES_DIR } from '../config';
 
 // Adaptive Micro-Practice & Spaced-Repetition: paper generation (single-
 // competency, multi-competency, and bulk), upload/grading, and the due/
 // progress/history views that drive the teacher-facing practice workflow.
+
+// execFileSync throws before we can read stdout normally, so pull the real
+// JSON error line our Python scripts print instead of its generic wrapper text.
+function extractPythonScriptError(e: any, fallbackPrefix: string): string {
+  const stdout: string = typeof e?.stdout === 'string' ? e.stdout : (e?.stdout ? String(e.stdout) : '');
+  const lastLine = stdout.trim().split('\n').filter(Boolean).pop();
+  if (lastLine) {
+    try {
+      const parsed = JSON.parse(lastLine);
+      if (parsed && parsed.error) return parsed.error;
+    } catch {
+      // not JSON — fall through to the generic message
+    }
+  }
+  return `${fallbackPrefix}: ${e?.message || e}`;
+}
 
 // Shared logic for weak-competency lookup — used by both
 // GET /api/students/:id/weak-competencies (single-student UI lookup) and
@@ -432,6 +448,105 @@ export function registerMicroPracticeRoutes(app: express.Express) {
 
     await dbStore.updateUploadedPaper(uploadedPaper.id, { draftAnswers: answers });
     res.json({ success: true });
+  });
+
+  // Rasterizes an uploaded PDF to PNG so jsQR (pixel-only) can read it;
+  // skips the blue-ink filter, which would destroy a printed QR. allPages renders every page for multi-page scans.
+  app.post('/api/icr/rasterize-pdf', async (req, res) => {
+    const user = getAuthUser(req);
+    if (!user) return res.status(401).json({ error: 'Unauthorized' });
+
+    const { fileBase64, allPages } = req.body || {};
+    if (!fileBase64 || typeof fileBase64 !== 'string') {
+      return res.status(400).json({ error: 'fileBase64 is required.' });
+    }
+    const pdfMatch = /^data:application\/pdf;base64,(.+)$/.exec(fileBase64);
+    if (!pdfMatch) {
+      return res.status(400).json({ error: 'fileBase64 must be a base64-encoded PDF data URL.' });
+    }
+    const buf = Buffer.from(pdfMatch[1], 'base64');
+    if (buf.length === 0) {
+      return res.status(400).json({ error: 'fileBase64 decoded to zero bytes.' });
+    }
+    if (buf.length > 8 * 1024 * 1024) {
+      return res.status(413).json({ error: 'File too large (max 8 MB).' });
+    }
+
+    const tempDir = path.join(AI_SERVICES_DIR, 'scratch');
+    fs.mkdirSync(tempDir, { recursive: true });
+    const stamp = Date.now() + '_' + randomUUID().slice(0, 8);
+    const inputPath = path.join(tempDir, `rasterize_${stamp}_in.pdf`);
+    const outputPath = path.join(tempDir, `rasterize_${stamp}_out.png`);
+    // pdf_rasterize.py's --all-pages mode treats its output argument as a
+    // DIRECTORY (writes page_1.png, page_2.png, ... inside it) — a distinct
+    // path from outputPath's single-file usage in the non-allPages case.
+    const pagesDir = path.join(tempDir, `rasterize_${stamp}_pages`);
+    const pageOutputPaths: string[] = [];
+
+    try {
+      fs.writeFileSync(inputPath, buf);
+
+      const rasterScript = path.join(AI_SERVICES_DIR, 'scripts', 'pdf_rasterize.py');
+      const { execFileSync } = await import('child_process');
+      const scriptArgs = [rasterScript, inputPath, allPages ? pagesDir : outputPath];
+      if (allPages) scriptArgs.push('--all-pages');
+      let stdout: string;
+      try {
+        stdout = execFileSync(
+          PYTHON_BIN,
+          scriptArgs,
+          { cwd: AI_SERVICES_DIR, timeout: 30000, encoding: 'utf8' }
+        );
+      } catch (e: any) {
+        return res.status(500).json({ success: false, error: extractPythonScriptError(e, 'PDF rasterization failed') });
+      }
+      const jsonLine = stdout.trim().split('\n').filter(Boolean).pop() || '{}';
+      let parsed: any = {};
+      try {
+        parsed = JSON.parse(jsonLine);
+      } catch {
+        return res.status(500).json({ success: false, error: `PDF rasterizer returned non-JSON: ${stdout.slice(0, 300)}` });
+      }
+      if (!parsed.success) {
+        return res.status(500).json({ success: false, error: parsed.error || 'PDF rasterization failed.' });
+      }
+
+      if (allPages) {
+        const pages: Array<{ output_path: string; page_number: number }> = parsed.pages || [];
+        const imageDataUrls = pages.map(p => {
+          pageOutputPaths.push(p.output_path);
+          const imgBuf = fs.readFileSync(p.output_path);
+          // pdf_rasterize.py's --all-pages mode writes JPEG (page_N.jpg); the
+          // declared MIME here must track its actual output extension, not be
+          // assumed, so a future format change on that side can't silently
+          // mislabel bytes again.
+          const ext = path.extname(p.output_path).toLowerCase();
+          const mimeType = ext === '.png' ? 'image/png' : 'image/jpeg';
+          return `data:${mimeType};base64,${imgBuf.toString('base64')}`;
+        });
+        return res.json({
+          success: true,
+          imageDataUrls,
+          pageCount: pages.length
+        });
+      }
+
+      const pngBuf = fs.readFileSync(outputPath);
+      res.json({
+        success: true,
+        imageDataUrl: `data:image/png;base64,${pngBuf.toString('base64')}`,
+        pageSize: parsed.page_size
+      });
+    } catch (err: any) {
+      const msg = err?.message || String(err);
+      console.error('[icr-rasterize-pdf] failed:', msg);
+      res.status(500).json({ success: false, error: msg });
+    } finally {
+      try { fs.unlinkSync(inputPath); } catch { /* noop */ }
+      try { fs.unlinkSync(outputPath); } catch { /* noop */ }
+      for (const p of pageOutputPaths) { try { fs.unlinkSync(p); } catch { /* noop */ } }
+      try { fs.rmSync(pagesDir, { recursive: true, force: true }); } catch { /* noop */ }
+    }
   });
 
   // Pre-flight check for client-decoded paperIds: reports which already have
